// frontend/src/types/Store.ts
// export interface StoreType {
//   id: number;
//   district: string;
//   subdivision: string;
//   block: string;
//   storeCode: string;
//   storeName: string;
//   storeType: string;
// }
export interface StoreType {
  id: number;

  district: string;

  subdivision: string;

  block: string;

  storeCode: string;

  storeName: string;

  storeType: string;

  storeIncharge: string;

  capacity: string;

  status: string;
}